refactor(block): rename block ID to UUID in block entry components
