+++ src/app/dashboard/events/[uuid]/blocks/[blockUuid]/block-participants-list.tsx
