+++ src/app/dashboard/events/[uuid]/blocks/[blockUuid]/toggle-participants-visibility-button.tsx
